+++ engine/hypersurface/nnormalhypersurface.cpp
@@ -96,13 +96,13 @@ void NNormalHypersurface::writeXMLData(std::ostream& out) const {
     using regina::xml::xmlValueTag;
 
     // Write the opening tag including vector length.
-    unsigned vecLen = vector_->size();
+    size_t vecLen = vector_->size();
     out << "  <hypersurface len=\"" << vecLen << "\" name=\""
         << xmlEncodeSpecialChars(name_) << "\">";
 
     // Write all non-zero entries.
     NLargeInteger entry;
-    for (unsigned i = 0; i < vecLen; i++) {
+    for (size_t i = 0; i < vecLen; i++) {
         entry = (*vector_)[i];
         if (entry != 0)
             out << ' ' << i << ' ' << entry;
